+++ schema.ts
@@ -55,6 +55,10 @@ export enum ValueHintType {
    */
   Duration = 'duration',
   /**
+   * Indicates to interpret the value as an email address (e.g. joe@foo.com).
+   */
+  Email = 'email',
+  /**
    * Indicates to interpret and render the value as a Coda person reference. The provided value should be
    * an object whose `id` property is an email address, which Coda will try to resolve to a user
    * and render an @-reference to the user.
@@ -159,6 +163,7 @@ export const StringHintValueTypes = [
   ValueHintType.Time,
   ValueHintType.DateTime,
   ValueHintType.Duration,
+  ValueHintType.Email,
   ValueHintType.Embed,
   ValueHintType.Html,
   ValueHintType.ImageReference,
@@ -393,6 +398,84 @@ export interface ScaleSchema extends BaseNumberSchema<ValueHintType.Scale> {
 }
 
 /**
+ * Display types that can be used with an {@link EmailSchema}.
+ */
+export enum EmailDisplayType {
+  /**
+   * Display both icon and email (default).
+   */
+  IconAndEmail = 'iconAndEmail',
+
+  /**
+   * Display icon only.
+   */
+  IconOnly = 'iconOnly',
+
+  /**
+   * Display email address only.
+   */
+  EmailOnly = 'emailOnly',
+}
+
+/**
+ * A schema representing a return value or object property that is an email address.
+ *
+ * An email address can be represented visually as an icon, an icon plus the email address, or
+ * the just the email address.  Autocomplete against previously typed domain names is
+ * also an option in the user interface.
+ */
+export interface EmailSchema extends BaseStringSchema<ValueHintType.Email> {
+  /** Instructs Coda to render this value as an email address. */
+  codaType: ValueHintType.Email;
+  /** How the email should be displayed in the UI. */
+  display?: EmailDisplayType;
+  /** Whether to auto-complete the email domain during user input. */
+  autocomplete?: boolean;
+}
+
+/**
+ * Display types that can be used with a {@link LinkSchema}.
+ */
+export enum LinkDisplayType {
+  /**
+   * Display icon only.
+   */
+  IconOnly = 'iconOnly',
+
+  /**
+   * Display URL.
+   */
+  Url = 'url',
+
+  /**
+   * Display web page title.
+   */
+  Title = 'title',
+
+  /**
+   * Display the referenced web page as a card.
+   */
+  Card = 'card',
+
+  /**
+   * Display the referenced web page as an embed.
+   */
+  Embed = 'embed',
+}
+
+/**
+ * A schema representing a return value or object property that is a hyperlink.
+ *
+ * The link can be displayed in the UI in multiple ways, as per the above enumeration.
+ */
+export interface LinkSchema extends BaseStringSchema<ValueHintType.Url> {
+  /** Instructs Coda to render this value as a hyperlink. */
+  codaType: ValueHintType.Url;
+  /** How the URL should be displayed in the UI. */
+  display?: LinkDisplayType;
+}
+
+/**
  * A schema representing a return value or object property that is provided as a string,
  * which Coda should interpret as a date. Coda is able to flexibly parse a number of formal
  * and informal string representations of dates. For maximum accuracy, consider using an
@@ -527,6 +610,7 @@ export const SimpleStringHintValueTypes = [
   ValueHintType.ImageAttachment,
   ValueHintType.Markdown,
   ValueHintType.Url,
+  ValueHintType.Email,
 ] as const;
 export type SimpleStringHintTypes = typeof SimpleStringHintValueTypes[number];
 
@@ -545,6 +629,8 @@ export type StringSchema =
   | StringTimeSchema
   | StringDateTimeSchema
   | DurationSchema
+  | EmailSchema
+  | LinkSchema
   | StringEmbedSchema
   | SimpleStringSchema;